crafting system implementation

+++ src/hooks/use-delete-edge.ts
@@ -1,7 +1,8 @@
 import { useCallback } from "react";
 import { Edge, useReactFlow } from "reactflow";
 import { useDispatch } from "react-redux";
-import { removeConnection } from "~/slices/groups/groups.ts";
+import { removeConnection } from "~/slices/resource-groups.ts";
+import { RESOURCE } from "~/constants/dictionary.ts";
 
 export function useDeleteEdge() {
   const { setEdges, getNode } = useReactFlow();
@@ -12,7 +13,7 @@ export function useDeleteEdge() {
       const source = getNode(edge.source);
       const target = getNode(edge.target);
 
-      if (source && target) {
+      if (source && target && source.data.roles.includes(RESOURCE) === target.data.roles.includes(RESOURCE)) {
         dispatch(removeConnection({ source, target }));
       }
